+++ spinta/datasets/backends/dataframe/backends/soap/components.py
@@ -16,3 +16,8 @@ class Soap(DaskBackend):
     @contextlib.contextmanager
     def begin(self):
         yield
+
+    # This backend is used as a parameter for _get_data_soap, which is then passed to Dask reader.
+    # Dask requires all the given parameters to be hashable, it uses __dask_tokenize__ to achieve it.
+    def __dask_tokenize__(self):
+        return self.type, self.name