+++ spinta/testing/pytest.py
@@ -148,7 +148,10 @@ def app(context) -> TestClient:
 
 @pytest.fixture
 def cli(rc: RawConfig):
-    yield SpintaCliRunner(mix_stderr=False)
+    try:
+        yield SpintaCliRunner(mix_stderr=False)
+    except Exception:
+        yield SpintaCliRunner()
     _remove_push_state(rc)