+++ CHANGES.rst
@@ -4,6 +4,13 @@ Changes
 0.2dev6 (unreleased)
 ====================
 
+Improvements:
+
+- Updated `authlib` minimal version to 1.0.0 (`#675`_).
+- `private` and `public` keys now include `kid` field (`#675`_).
+
+  .. _#675: https://github.com/atviriduomenys/spinta/issues/675
+
 - Removed dependency `mypy`
 
 0.2dev5 (2025-09-03)